rename config names
