+++ feathr_project/feathr/definition/settings.py
@@ -1,7 +1,7 @@
 from typing import Optional
 from jinja2 import Template
 from loguru import logger
-from feathr.frameconfig import HoconConvertible
+from feathr.definition.feathrconfig import HoconConvertible
 
 class ObservationSettings(HoconConvertible):
     """Time settings of the observation data. Used in feature join.
